% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_all_posts}
\alias{build_all_posts}
\title{Builds all blog posts}
\usage{
build_all_posts()
}
\description{
Builds all drafted blog posts for a DACSS course site.
The index page listing posts will be updated once
the Distill blog is built. Assumes a DACSS blog R project
is being used.
}
