% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_lates_posts}
\alias{build_lates_posts}
\title{Build latest blog posts}
\usage{
build_lates_posts()
}
\description{
Render the posts added or modified in the most recent git pull.
Recommended use is only after git pull. In other cases and for a foolproof rendering of posts, consider using build_all_posts().
}
