% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_student_pages}
\alias{build_student_pages}
\title{Builds all student pages}
\usage{
build_student_pages()
}
\description{
Builds all student pages for a DACSS course site.
Assumes the working directory is the GitHub repo
project file.
}
