% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude_dir}
\alias{exclude_dir}
\title{Exclude directories in your repo}
\usage{
exclude_dir(path)
}
\arguments{
\item{path}{The path to the directory you want to exclude}
}
\description{
Exclude directories in your GitHub repo
}
