% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude_docs}
\alias{exclude_docs}
\title{Exclude the docs folder}
\usage{
exclude_docs()
}
\description{
Exclude the docs folder. Students will need to use
this function to ensure they do not commit
or push the docs folder to their forked repos.
}
