% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude_file}
\alias{exclude_file}
\title{Exclude files in your repo}
\usage{
exclude_file(path)
}
\arguments{
\item{path}{The path to the file you want to exclude}
}
\description{
Exclude files in your GitHub repo.
}
