% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_dacss_blog}
\alias{generate_dacss_blog}
\title{Generate a new DACSS blog}
\usage{
generate_dacss_blog()
}
\description{
Generates a new DACSS blog into a user-specified GitHub
repository. A generic DACSS blog will be
created with a Distill and Postcards setup. Assumes the user
has Git configured on their personal computer already.
The repo will be generated if it has not been setup already.
Simply follow the series of prompts when the function is called.
}
