% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_student_pages}
\alias{generate_student_pages}
\title{Generate student postcards}
\usage{
generate_student_pages(spreadsheet, names_col, theme = "jolla", ...)
}
\arguments{
\item{spreadsheet}{A csv or xlsx file containing student names.}

\item{names_col}{The column containing student names. Can be the column
name or number.}

\item{theme}{The template theme to create postcard pages with. The default
value is "jolla".}

\item{...}{Additional arguments can be taken based on arguments for
read_csv or read_xlsx.}
}
\description{
Generates 'about me' pages for students in a Postcards setup. A spreadsheet
of student information is taken as input -- with names as a primary
requirement. Csv and xlsx formats are accepted.
It will also generate list of student names in the Student page.
Assumes the working directory is based on the GitHub blog repo prokect.
}
