% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_all_posts}
\alias{render_all_posts}
\title{Render all blog posts}
\usage{
render_all_posts()
}
\description{
Renders all drafted blog posts for a DACSS course site.
The index page listing posts will be updated once
the Distill blog is built. Assumes a DACSS blog R project
is being used.
}
