% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_student_pages}
\alias{render_student_pages}
\title{Renders all student pages}
\usage{
render_student_pages()
}
\description{
Renders all student pages for a DACSS course site.
Assumes the working directory is the GitHub repo
project file.
}
