% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset_project_env}
\alias{reset_project_env}
\title{Reset the Project Environment}
\usage{
reset_project_env()
}
\description{
Resets the project environment to the initial settings
created from the Google Spreadsheet inputs.
}
