% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_course_instructor}
\alias{update_course_instructor}
\title{Update Course Instructor name}
\usage{
update_course_instructor(name, prof_pic = NULL)
}
\arguments{
\item{name}{The name to replace the current instructor name with.}

\item{prof_pic}{To replace profile pic. Optional.}
}
\description{
Update Course Instructor name in the homepage.
}
