% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_course_repo}
\alias{update_course_repo}
\title{Update Course Repo Link}
\usage{
update_course_repo(link = NULL)
}
\arguments{
\item{link}{The new repo link to replace the current repo link with.}
}
\description{
Update Course Repo Link in the header.
}
