% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_course_semester}
\alias{update_course_semester}
\title{Update course semester}
\usage{
update_course_semester(semester)
}
\arguments{
\item{semester}{The semester to replace the current semester with.}
}
\description{
Update course semester in the homepage.
}
