% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_course_title}
\alias{update_course_title}
\title{Update course title}
\usage{
update_course_title(title)
}
\arguments{
\item{title}{The title to replace the current course title with.}
}
\description{
Update course title in the homepage.
}
