% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_course_blog}
\alias{create_course_blog}
\title{Builds a new DACSS blog}
\usage{
create_course_blog()
}
\description{
Builds a new DACSS blog into a user-specified GitHub
repository. A generic DACSS blog will be
created with a Distill and Postcards setup. Checks if the user
has Git configured on their personal computer. The repo will be generated
if it has not been setup already. Also asks the user to authenticate their
Google account. Simply follow the series of prompts when the function is called.
}
